scripts/requirements.txt
requests>=2.28.0
pyyaml>=6.0
pycryptodome>=3.19.0
